move drawer cache into storage cache module
